+++ src/classes/job.ts
@@ -44,6 +44,7 @@ const logger = debuglog('bull');
 const optsDecodeMap = {
   de: 'deduplication',
   fpof: 'failParentOnFailure',
+  cpof: 'continueParentOnFailure',
   idof: 'ignoreDependencyOnFailure',
   kl: 'keepLogs',
   rdof: 'removeDependencyOnFailure',
@@ -952,7 +953,7 @@ export class Job<
   }
 
   /**
-   * Get jobs children failure values that were ignored if any.
+   * Get job's children failure values that were ignored if any.
    *
    * @returns Object mapping children job keys with their failure values.
    */
@@ -961,7 +962,7 @@ export class Job<
   }
 
   /**
-   * Get jobs children failure values that were ignored if any.
+   * Get job's children failure values that were ignored if any.
    *
    * @deprecated This method is deprecated and will be removed in v6. Use getIgnoredChildrenValues instead.
    *
@@ -1295,6 +1296,13 @@ export class Job<
   }
 
   protected validateOptions(jobData: JobJson) {
+    const exclusiveOptions: (keyof JobsOptions)[] = [
+      'removeDependencyOnFailure',
+      'failParentOnFailure',
+      'continueParentOnFailure',
+      'ignoreDependencyOnFailure',
+    ];
+
     const exceedLimit =
       this.opts.sizeLimit &&
       lengthInUtf8Bytes(jobData.data) > this.opts.sizeLimit;
@@ -1309,24 +1317,14 @@ export class Job<
       throw new Error(`Delay and repeat options could not be used together`);
     }
 
-    if (this.opts.removeDependencyOnFailure && this.opts.failParentOnFailure) {
-      throw new Error(
-        `RemoveDependencyOnFailure and failParentOnFailure options can not be used together`,
-      );
-    }
+    const enabledExclusiveOptions = exclusiveOptions.filter(
+      opt => this.opts[opt],
+    );
 
-    if (
-      this.opts.removeDependencyOnFailure &&
-      this.opts.ignoreDependencyOnFailure
-    ) {
+    if (enabledExclusiveOptions.length > 1) {
+      const optionsList = enabledExclusiveOptions.join(', ');
       throw new Error(
-        `RemoveDependencyOnFailure and ignoreDependencyOnFailure options can not be used together`,
-      );
-    }
-
-    if (this.opts.failParentOnFailure && this.opts.ignoreDependencyOnFailure) {
-      throw new Error(
-        `FailParentOnFailure and ignoreDependencyOnFailure options can not be used together`,
+        `The following options cannot be used together: ${optionsList}`,
       );
     }