+++ src/classes/scripts.ts
@@ -185,7 +185,13 @@ export class Scripts {
     const queueKeys = this.queue.keys;
 
     const parent: Record<string, any> = job.parent
-      ? { ...job.parent, fpof: opts.fpof, rdof: opts.rdof, idof: opts.idof }
+      ? {
+          ...job.parent,
+          fpof: opts.fpof,
+          rdof: opts.rdof,
+          idof: opts.idof,
+          cpof: opts.cpof,
+        }
       : null;
 
     const args = [
@@ -653,6 +659,7 @@ export class Scripts {
           ? opts.metrics?.maxDataPoints
           : '',
         fpof: !!job.opts?.failParentOnFailure,
+        cpof: !!job.opts?.continueParentOnFailure,
         idof: !!job.opts?.ignoreDependencyOnFailure,
         rdof: !!job.opts?.removeDependencyOnFailure,
       }),