+++ tests/test_worker.ts
@@ -21,6 +21,8 @@ import {
   removeAllQueueData,
 } from '../src/utils';
 
+const NoopProc = () => Promise.resolve();
+
 describe('workers', function () {
   const redisHost = process.env.REDIS_HOST || 'localhost';
   const prefix = process.env.BULLMQ_TEST_PREFIX || 'bull';
@@ -316,7 +318,7 @@ describe('workers', function () {
 
     await Promise.all(jobs);
 
-    const worker = new Worker(queueName, async () => {}, {
+    const worker = new Worker(queueName, NoopProc, {
       connection,
       prefix,
     });
@@ -492,7 +494,7 @@ describe('workers', function () {
       async job => {
         expect(job.data.foo).to.be.equal('bar');
         if (job.data.index == 1) {
-          const job = await queue.add('test', { foo: 'bar', index: 2 });
+          await queue.add('test', { foo: 'bar', index: 2 });
           await delay(1000);
         }
         return 'my data';
@@ -563,8 +565,8 @@ describe('workers', function () {
     expect(bclientSpy.callCount).to.be.gte(0);
     expect(bclientSpy.callCount).to.be.lte(1);
 
-    await new Promise<void>((resolve, reject) => {
-      worker.on('completed', (_job: Job, _result: any) => {
+    await new Promise<void>(resolve => {
+      worker.on('completed', () => {
         completedJobs++;
         if (completedJobs == numJobs) {
           resolve();
@@ -609,8 +611,8 @@ describe('workers', function () {
 
     expect(bclientSpy.callCount).to.be.equal(0);
 
-    await new Promise<void>((resolve, reject) => {
-      worker.on('completed', (job: Job, result: any) => {
+    await new Promise<void>(resolve => {
+      worker.on('completed', () => {
         completedJobs++;
         if (completedJobs == numJobs) {
           resolve();
@@ -660,7 +662,7 @@ describe('workers', function () {
 
     await anotherWorker.waitUntilReady();
 
-    await new Promise<void>((resolve, reject) => {
+    await new Promise<void>(resolve => {
       worker.on('completed', async () => {
         resolve();
       });
@@ -897,7 +899,7 @@ describe('workers', function () {
 
   describe('when 0.002 is used as blocktimeout', () => {
     it('should not block forever', async () => {
-      const worker = new Worker(queueName, async () => {}, {
+      const worker = new Worker(queueName, NoopProc, {
         connection,
         prefix,
       });
@@ -997,7 +999,7 @@ describe('workers', function () {
     describe('when blockUntil is 0', () => {
       describe('when drainDelay is greater than minimumBlockTimeout', () => {
         it('returns drainDelay', async () => {
-          const worker = new Worker(queueName, async () => {}, {
+          const worker = new Worker(queueName, NoopProc, {
             connection,
             prefix,
             autorun: false,
@@ -1010,7 +1012,7 @@ describe('workers', function () {
 
       describe('when drainDelay is lower than minimumBlockTimeout', () => {
         it('returns drainDelay', async () => {
-          const worker = new Worker(queueName, async () => {}, {
+          const worker = new Worker(queueName, NoopProc, {
             connection,
             drainDelay: 0.00001,
             prefix,
@@ -1031,7 +1033,7 @@ describe('workers', function () {
     describe('when blockUntil is greater than 0', () => {
       describe('when blockUntil is lower than date now value', () => {
         it('returns blockDelay value lower or equal 0', async () => {
-          const worker = new Worker(queueName, async () => {}, {
+          const worker = new Worker(queueName, NoopProc, {
             connection,
             prefix,
             autorun: false,
@@ -1047,7 +1049,7 @@ describe('workers', function () {
 
       describe('when blockUntil is greater than date now value', () => {
         it('returns delay value greater than minimumBlockTimeout', async () => {
-          const worker = new Worker(queueName, async () => {}, {
+          const worker = new Worker(queueName, NoopProc, {
             connection,
             prefix,
             autorun: false,
@@ -1868,7 +1870,7 @@ describe('workers', function () {
     describe('when run method is called when worker is running', function () {
       it('throws error', async () => {
         const maxJobs = 10;
-        const worker = new Worker(queueName, async () => {}, {
+        const worker = new Worker(queueName, NoopProc, {
           autorun: false,
           connection,
           prefix,
@@ -2150,7 +2152,7 @@ describe('workers', function () {
     this.timeout(4000);
     expect(
       () =>
-        new Worker(queueName, async () => {}, {
+        new Worker(queueName, NoopProc, {
           connection,
           prefix,
           stalledInterval: 0,
@@ -2162,7 +2164,7 @@ describe('workers', function () {
     this.timeout(4000);
     expect(
       () =>
-        new Worker(queueName, async () => {}, {
+        new Worker(queueName, NoopProc, {
           connection,
           prefix,
           drainDelay: 0,
@@ -2207,7 +2209,7 @@ describe('workers', function () {
   describe('Concurrency process', () => {
     it('should thrown an exception if I specify a concurrency of 0', () => {
       try {
-        const worker = new Worker(queueName, async () => {}, {
+        const worker = new Worker(queueName, NoopProc, {
           connection,
           prefix,
           concurrency: 0,
@@ -2222,7 +2224,7 @@ describe('workers', function () {
 
     it('should thrown an exception if I specify a NaN concurrency', () => {
       try {
-        const worker = new Worker(queueName, async () => {}, {
+        const worker = new Worker(queueName, NoopProc, {
           connection,
           prefix,
           concurrency: NaN,
@@ -2736,7 +2738,7 @@ describe('workers', function () {
 
     describe('when jobs do not fail and get the maximum attempts limit', () => {
       it('does not emit retries-exhausted event', async () => {
-        const worker = new Worker(queueName, async () => {}, {
+        const worker = new Worker(queueName, NoopProc, {
           connection,
           prefix,
         });
@@ -4586,7 +4588,7 @@ describe('workers', function () {
   });
 
   it('should retrieve concurrency from getter', async () => {
-    const worker = new Worker(queueName, async () => {}, {
+    const worker = new Worker(queueName, NoopProc, {
       connection,
       concurrency: 100,
     });